+++ src/Public/Tags/Get-FrontCompanyTags.ps1
@@ -1,0 +1,35 @@
+function Get-FrontCompanyTags {
+    <#
+    .SYNOPSIS
+        Short description
+    .DESCRIPTION
+        Long description
+    .EXAMPLE
+        PS C:\> <example usage>
+        Explanation of what the example does
+    .INPUTS
+        Inputs (if any)
+    .OUTPUTS
+        Output (if any)
+    #>
+    param (
+        [Parameter()]
+        [String]$Id,
+
+        [Parameter(Mandatory)]
+        [SecureString]$ApiKey
+    )
+
+    $Params = @{
+        Method   = "GET"
+        Endpoint = "company/tags"
+        ApiKey   = $ApiKey
+    }
+
+    try {
+        InvokeFrontRestMethod @Params
+    }
+    catch {
+        Write-Error -ErrorRecord $_
+    }
+}